Fixes to build on Linux

+++ src/dynd/dtypes/array_dtype.cpp
@@ -61,7 +61,7 @@ dtype dynd::array_dtype::apply_linear_index(int nindices, const irange *indices,
 void dynd::array_dtype::get_shape(int i, std::vector<intptr_t>& out_shape) const
 {
     // Ensure the output shape is big enough
-    while (out_shape.size() <= i) {
+    while (out_shape.size() <= (size_t)i) {
         out_shape.push_back(shape_signal_uninitialized);
     }
